addressed issues with test harness
