+++ packages/teraslice-op-test-harness/utils.js
@@ -1,0 +1,16 @@
+'use strict';
+
+function bindThis(instance, cls) {
+    return Object.getOwnPropertyNames(Object.getPrototypeOf(instance))
+        .filter((name) => {
+            const method = instance[name];
+            return method instanceof Function && method !== cls;
+        })
+        .forEach((mtd) => {
+            instance[mtd] = instance[mtd].bind(instance);
+        });
+}
+
+module.exports = {
+    bindThis
+};
